✨ feat(attrs): backfill type annotations for attrs classes (#181)
Classic `@attr.s` with `attr.ib(type=...)` doesn't set `__annotations__`,
so Sphinx can't discover attribute types. This adds a backfill step that
reads types from attrs field metadata before docstring processing.

Split `_resolver.py` (454 lines) into a `_resolver/` package with focused
submodules: `_type_hints`, `_type_comments`, `_stubs`, `_attrs`, and `_util`.
The package `__init__.py` re-exports all public symbols so existing imports
continue to work unchanged. Test suite reorganized to mirror the new structure.
