+++ pyproject.toml
@@ -56,6 +56,7 @@ dev = [
   { include-group = "type" },
 ]
 test = [
+  "attrs>=22",
   "covdefaults>=2.3",
   "coverage>=7.13.4",
   "defusedxml>=0.7.1",
@@ -133,7 +134,7 @@ lint.per-file-ignores."tests/roots/test-pyi-stubs/stub_mod.py" = [
   "PLR6301", # method must be instance method to test stub resolution
   "RUF029",  # async function must be async to test async stub resolution
 ]
-lint.per-file-ignores."tests/test_resolver.py" = [
+lint.per-file-ignores."tests/test_resolver/test_type_comments.py" = [
   "ANN001", # type comment fixtures intentionally omit annotations
   "ARG001", # type comment fixtures have intentionally unused args
 ]
