+++ tests/roots/test-attrs/index.rst
@@ -1,0 +1,2 @@
+Test
+====
